Trebuchet: Porting of Operator home screen customization.
Added different custom default_workspace.xml which contains
reordering of hotseat items and added new hot seat items.
Google search view is removed and google search widget is added
only to middle screen.
Feature control flag is added.Based on the flag these
features will be loaded in Launcher.Flag is set to false in
Launcher and overlayed to true in Regional/Latam folder.
When LATAM region setup will be loaded the flag will be set
to true and the above customization will be loaded.

Change-Id: Iaf8496c29804fe4a2a2f4a270c4ccb4460201110
CRs-Fixed: 797995