Trebuchet: Porting of operator home screen customization.
Added different custom default_workspace.xml which contains
reordering of hotseat items and added new hot seat items.
Google search view is removed and google search widget is added
only to middle screen.
Feature control flag is added.Based on the flag these
features will be loaded in Launcher.Flag is set to false in
Launcher and overlayed to true in Regional/Latam folder.
When required region setup will be loaded the flag will be set
to true and the above customization will be loaded.

Change-Id: I9f82ff522d2331ae9ff6c013c4c81d92ad5266ba
CRs-Fixed: 797995