Trebuchet: Porting of Operator home screen customization.
Added different custom default_workspace.xml which contains
reordering of hotseat items and added new hot seat items.
Google search view is removed and google search widget is added
only to middle screen.
Feature control flag is added.Based on the flag these
features will be loaded in Launcher.Flag is set to false in
Launcher and overlayed to true in Regional/Latam folder.
When specific country setup will be loaded the flag will be set
to true and the above customization will be loaded.

Change-Id: I24833eb6dc2e1673ef9b362e4731197bd6003a63
CRs-Fixed: 799351